% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_per_base_seq_qual.R
\name{plot_per_base_seq_qual}
\alias{plot_per_base_seq_qual}
\title{Function for plotting Per Base Sequencing Quality from FastQC}
\usage{
plot_per_base_seq_qual(df, yaxis = NULL, color = NULL, facet_by = NULL)
}
\arguments{
\item{df}{Dataframe as shown in the vignette.}

\item{yaxis}{ggplot2 compatible symbol relating to a column in df. Passed
to geom_point to plot the y-axis variable.}

\item{color}{ggplot2 compatible symbol relating to a column in df. Passed
to geom_point to the color parameter.}

\item{facet_by}{ggplot2 compatible symbol relating to a column in df. Passed
to facet_wrap.}
}
\value{
ggplot plot of the data (invisibly).
}
\description{
This functions plots the values provided from the \link{reading_in} function.
}
\examples{
fastqc_plotr_sample_df \%>\%
   plot_per_base_seq_qual()


}
